feat: add TCP/IP network transport, CLI stub, and cleanup
- Add src/transport/net.rs with NetConnection implementing
  Connection trait (client/server, stats, drain_input)
- Add PilotSocket::net_listen() and accept() for server mode
- Implement Connection trait for Serial and Usb
- Add src/main.rs CLI stub with clap (sync, list, info, dump)
- Rename Connection::flush to drain_input to avoid Write::flush
  collision
- Fix partial-write progress loss in NetConnection::write
- Fix PilotSocket::net_listen() to call conn.listen()
- Add SAFETY comments for unsafe Usb code
- Cleanup: derive(Default), .get(0)->.first(), unused params
  across 20+ files
- Add lint suppressions in lib.rs for pre-existing warnings

+++ src/records/todo.rs
@@ -142,7 +142,7 @@ impl TodoRecord {
         } else if let Some(ref due) = self.due {
             let (year, month, day) = due.get_date();
             // Format: YYYYYYYMMMMMDDDDD
-            let due_short = (((year as u16 - 4) & 0x7F) << 9) |
+            let due_short = (((year - 4) & 0x7F) << 9) |
                            (((month as u16 + 1) & 0x0F) << 5) |
                            ((day as u16) & 0x1F);
             data.extend_from_slice(&due_short.to_le_bytes());
